… tests/ImageSharp.Tests/Formats/Jpg/DCTTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/Utils/JpegFixture.cs
… tests/ImageSharp.Tests/Formats/Jpg/Utils/LibJpegTools.ComponentData.cs
… tests/ImageSharp.Tests/Formats/Jpg/Utils/ReferenceImplementations.cs
… tests/ImageSharp.Tests/Formats/Jpg/ZigZagTests.cs